### 2021-06-14 #3366 @whedon
[PRE REVIEW]: OGRe: An Object-Oriented General Relativity Package for Mathematica
**Submitting author:** @bshoshany (<a href="http://orcid.org/0000-0003-2222-127X">Barak Shoshany</a>)
**Repository:** <a href="https://github.com/bshoshany/OGRe" target ="_blank">https://github.com/bshoshany/OGRe</a>
**Version:** v1.5
