### 2021-06-14 #3366 @bshoshany
I went over the list of people who have already agreed to review for JOSS. The following potential reviewers listed Mathematica as one of their programming languages and wrote astronomy as one of their domains (so they should have knowledge of general relativity):
1. danehkar
2. kostunin
3. bradkav
4. eschnett (note: I have worked with him in the past, so that may be a conflict of interest)
