### 2021-06-24 #3366 @VivianePons
Thank you @kostunin . Once we launch the review, you will get a checklist to help you go through. If ever there is an item where you don't feel competent, you just let me know and we'll decide how to handle it.
