### 2021-07-27 #3416 @amelialdrew
Hi @bshoshany, this is looking good to me so far, it seems as though it would be useful for students getting to grips with Mathematica/GR and the documentation is very clear. I have a few small questions/comments organised roughly into the checkpoints I haven’t ticked off yet:

- Do you have any examples of where the code has been cited/used for more advanced research projects?

- Main parallelisation section for me only speeds up by 30-40% and I get an error when running `TSetParallelization[True]`:

```
Parallelization enabled.
Part: Part 1 of {} does not exist
```

Other performance/optimisation claims throughout e.g. ‘Built with speed and performance in mind using optimized algorithms’ are not backed up with numbers, but this may not be necessary.

- Target audience is in Summary rather than Statement of Need, I expect this is fine.

- I think it would be worth citing the references throughout rather than just the list at the end. I would also suggest including a review of other GR/tensor manipulation codes, e.g. https://link.springer.com/article/10.1007/s41114-018-0015-6. Also, the link for the first review on the paper doesn’t work.

Willing to discuss all comments as I am not yet an experienced reviewer.
