### 2021-07-28 #3416 @bshoshany
Hi @amelialdrew,

Thanks so much for reviewing my package and for your very helpful comments! Here are my replies:

* Unfortunately I do not have any examples of this code being used in research yet, except for my own research. The package is only a few months old and it will probably take some time before people who are using it produce any papers with it. However, it is currently being used by both myself and 4 of my students, and I hope to publish the first paper citing the package in a couple of months.
* How many cores does your system have? The speedup in the documentation is with 12 cores and 8 kernels. If you have less than that, then the speedup will be lower.
* When `TSetParallelization` is turned on, it tries to start all available kernels, but executing `LaunchKernels[]` would fail if some kernels are already running, so it first determines how many kernels are already running and then starts only those that are not. You may have found a bug. When you execute `$ConfiguredKernels`, what do you see? It should be something like `{<<N local kernels>>}` and `$ConfiguredKernels[[1, 1]]` should return the total number of kernels available. From your error message I suspect `$ConfiguredKernels` does not return a `List` on your system. Which version of Mathematica are you using?
* The speedup from the parallelization is probably the only one that can be directly measured. But when I say I built it with speed and performance in mind, I am also referring to many other things happening behind the scene. For example:
1. The components of the tensor in any particular representation are not calculated until they are needed. Once they are calculated, they are stored permanently within the tensor object so they don't need to be calculated again later.
2. If you're trying to calculate e.g. the Ricci tensor, then OGRe first checks if the Riemann tensor has already been calculated. If so, then it uses the result of that calculation when calculating the Ricci tensor, instead of calculating it from scratch. And similarly for all other curvature tensors.
3. The ability to export and import tensor objects means that in your next Mathematica session you can import the tensors with everything that has been calculated so far (components in different representations, curvature tensors, etc.) instead of having to calculate everything from scratch.
4. And so on.
* If you want, I'm happy to move the target audience from the Summary to the Statement of Need.
* The references I linked to are just general references for differential geometry and general relativity that don't necessarily apply to any specific word or sentence, but to the paper as a whole. However, I can convert them to inline references if you prefer.
* I'm not sure why the ISBN link to the Baez and Munian book isn't working. The ISBN is correct but the link itself was generated by JOSS. I'll replace it with a different ISBN (9789810217297) for which the link does seem to work.
* I'll add a reference to the paper by MacCallum, which indeed seems relevant.

Let me know which of the above changes you want me to make, and then I'll upload a new version of the paper!
