### 2021-07-30 #3416 @amelialdrew
Hi @bshoshany,

Thank you for the quick reply and for the further information regarding parallelisation. My system has 4 cores which partially explains the slower speedup, but I would say not entirely. Running ```$ConfiguredKernels``` I get just ```{}```, so it seems you are right that it is not returning a ```List```. I am using Mathematica 12.3.1. Do you need any further info to help with the bug?

Further to sorting out this parallelisation issue, I would recommend:
- Moving target audience to Statement of Need (for no other reason than it says that specifically on our checklist!)
- Converting some of the more specific references to inline e.g. xACT package, adding the review paper as discussed and fixing the link.

Let me know what you think!
