### 2021-07-30 #3416 @bshoshany
@amelialdrew, are you running the code on the Wolfram Cloud, by any chance? That might explain why `$ConfiguredKernels` returns `{}`. If not, how many kernels does your Mathematica license allow? This can be found under Preferences -> Local Kernels. You will see an option "Limit by license availability (N)" where N is the number of kernels available in your license. And which OS are you using?

In any case, I will fix this bug by simply not launching any kernels in advance if `$ConfiguredKernels` is empty. The kernels will then instead be launched automatically by Mathematica the first time a parallelized task is executed. I will incorporate this fix into the next release, which should be ready within a week, and will also include some cool new features I've been working on, such as calculating the geodesic equation, more compact display of tensor components using `TList` and `TShow`, and more.

I will make the changes to `paper.md` as you requested, and upload it to the `paper` branch when I'm done. I will take that opportunity to also add the additional features from the upcoming v1.6 release to the list of features in the paper. I'll do that once the new release is up.
