### 2021-07-30 #3416 @amelialdrew
@bshoshany No I am not, although I've just looked under Preferences and it says ```Local Kernels is deprecated```, so I assume this more up to date version of Mathematica has removed them. I am using macOS Big Sur 11.4. The updates sound good!
