### 2021-07-30 #3416 @bshoshany
Hmm... I also use v12.3.1, and I also see the message "Local Kernels is deprecated", but I also see the number of local kernels in the preferences, and my `$ConfiguredKernels` returns a `List`. I get that on both Windows and Linux (I don't have access to a Mac).

The documentation for `$ConfiguredKernels` says that "As of Version 12.3, `$ConfiguredKernels` has been superseded by `$DefaultKernels`", but since I want the package to be compatible with Mathematica all the way down to version 12.0 (some people do not upgrade because that costs extra), I can't use that.

However, I found a simple solution; `CloseKernels[]; LaunchKernels[]` should launch all the available kernels regardless of how many are already running, so I'll just do that instead of what I do now for `TSetParallelization`.

I can't find any information in the release notes or documentation about the Local Kernels themselves being deprecated. Perhaps there is some new parallelization framework that doesn't use multiple local kernels, but I can find any evidence of that. I might have to contact Wolfram directly to ask what the deprecation means.

I have a few more questions to understand the bug better, if you don't mind:

1. What do yet get when you execute `$MaxLicenseSubprocesses`?
2. Start a fresh Mathematica session (after closing all the currently open Mathematica windows if any) and type `LaunchKernels[]`. The output should be a list of kernels. How many kernels are in the list?

Thanks!
