### 2021-08-02 #3416 @amelialdrew
Ah I've just realised there's an option for ```Enable Local Kernels```  which by default runs with 4 kernels with 4 CPUs which approximately halves the time vs no parallelisation, so this may fix it. I suppose it would be worth adding a line in the docs to make sure they are enabled.

In case it is still needed, ```$MaxLicenseSubprocesses``` gives me ```16```, and ```LaunchKernels[]``` gives 4 kernels, both after switching on ```Enable Local Kernels```.

I hope this helps!
