+++ OGRe.m
@@ -41,6 +41,9 @@ Null[{
     TCalc,
     TCalcChristoffel,
     TCalcEinsteinTensor,
+    TCalcGeodesicFromChristoffel,
+    TCalcGeodesicFromLagrangian,
+    TCalcLagrangian,
     TCalcRicciScalar,
     TCalcRicciTensor,
     TCalcRiemannTensor,
@@ -58,19 +61,23 @@ Null[{
     TImport,
     TImportAll,
     TInfo,
-    TInitializeSymbols,
     TLineElement,
     TList,
+    TMessage,
     TNewCoordinates,
     TNewMetric,
     TNewTensor,
     TPartialD,
+    TSetAllowOverwrite,
     TSetAssumptions,
     TSetAutoUpdates,
+    TSetCurveParameter,
     TSetIndexLetters,
     TSetParallelization,
+    TSetReservedSymbols,
     TShow,
-    TSimplify
+    TSimplify,
+    TVolumeElementSquared
 }];
 
 
@@ -78,7 +85,7 @@ Begin["`Private`"]; (* OGRe`Private` *)
 
 
 (* DO NOT change the format of the next line. It is used by TCheckForUpdates to detect the version of this file. Changing it will break the automatic update mechanism. Only change the version number and date. *)
-OGReVersion = "v1.5 (2021-06-07)";
+OGReVersion = "v1.6 (2021-08-07)";
 
 
 (* The raw URL of this file on GitHub. *)
@@ -95,11 +102,17 @@ If[
 OGReGlobalOptions = LocalSymbol["OGReGlobalOptions"];
 (* Set the "OGReVersion" key to the current version of the package. *)
 OGReGlobalOptions["OGReVersion"] = OGReVersion;
-(* If the option "AutoUpdates" has not been set, or is not a boolean value, we set it now to True. *)
+(* If the option "AutoUpdates" has not been set, or is not a boolean value, we set it now to the default value of True. *)
 If[
     !KeyExistsQ[OGReGlobalOptions, "AutoUpdates"] || !BooleanQ[OGReGlobalOptions["AutoUpdates"]],
 (* Then *)
     OGReGlobalOptions["AutoUpdates"] = True;
+];
+(* If the option "AllowOverwrite" has not been set, or is not a boolean value, we set it now to the default value of False. *)
+If[
+    !KeyExistsQ[OGReGlobalOptions, "AllowOverwrite"] || !BooleanQ[OGReGlobalOptions["AllowOverwrite"]],
+(* Then *)
+    OGReGlobalOptions["AllowOverwrite"] = False;
 ];
 (* Save the global options to the persistent storage. *)
 LocalSymbol["OGReGlobalOptions"] = OGReGlobalOptions;
@@ -352,8 +365,8 @@ TCalcChristoffel[metricID_String] := Module[
         christoffelID,
         inverseMetricID = NewTempID[]
     },
-    (* Check that the tensor does not already exist. *)
-    CheckIfOverwriting[metricID <> "Christoffel"];
+    (* Check that metricID exists. *)
+    CheckIfTensorExists[metricID];
     (* Check that metricID is indeed a metric. *)
     If[
         TensorData[metricID]["Role"] =!= "Metric",
@@ -392,8 +405,8 @@ TCalcEinsteinTensor[metricID_String] := Module[
     {
         EinsteinID
     },
-    (* Check that the tensor does not already exist. *)
-    CheckIfOverwriting[metricID <> "Einstein"];
+    (* Check that metricID exists. *)
+    CheckIfTensorExists[metricID];
     (* Check that metricID is indeed a metric. *)
     If[
         TensorData[metricID]["Role"] =!= "Metric",
@@ -419,14 +432,245 @@ TCalcEinsteinTensor[metricID_String] := Module[
 ];
 
 
[…]
