+++ README.md
@@ -5,6 +5,7 @@
 ![GitHub last commit](https://img.shields.io/github/last-commit/bshoshany/OGRe)
 [![GitHub repo stars](https://img.shields.io/github/stars/bshoshany/OGRe?style=social)](https://github.com/bshoshany/OGRe)
 [![Twitter @BarakShoshany](https://img.shields.io/twitter/follow/BarakShoshany?style=social)](https://twitter.com/BarakShoshany)
+[![Open in Visual Studio Code](https://open.vscode.dev/badges/open-in-vscode.svg)](https://open.vscode.dev/bshoshany/OGRe)
 
 # OGRe: An Object-Oriented General Relativity Package for Mathematica
 
@@ -14,22 +15,18 @@ Department of Physics, Brock University,\
 [bshoshany@brocku.ca](mailto:bshoshany@brocku.ca) | [https://baraksh.com/](https://baraksh.com/)\
 DOI: [doi:10.5281/zenodo.4742935](https://doi.org/10.5281/zenodo.4742935)
 
-<!-- TOC depthFrom:2 -->
+* [Summary](#summary)
+* [Features](#features)
+* [Installing and loading the package](#installing-and-loading-the-package)
+* [Documentation](#documentation)
+* [Issue and pull request policy](#issue-and-pull-request-policy)
+* [Copyright and citing](#copyright-and-citing)
 
-- [Summary](#summary)
-- [Features](#features)
-- [Installing and loading the package](#installing-and-loading-the-package)
-- [Documentation](#documentation)
-- [Copyright and citing](#copyright-and-citing)
-
-<!-- /TOC -->
-
-<a id="markdown-summary" name="summary"></a>
 ## Summary
 
-OGRe is a modern Mathematica package for tensor calculus, designed to be both powerful and user-friendly. It can be used in a variety of contexts where tensor calculations are needed, in both mathematics and physics, but it is especially suitable for general relativity.
+OGRe is a modern Mathematica package for differential geometry and tensor calculus, designed to be both powerful and user-friendly. It can be used in a variety of contexts where tensor calculations are needed, in both mathematics and physics, but it is especially suitable for general relativity.
 
-Tensors are abstract objects, which can be represented as multi-dimensional arrays once a choice of index configuration and coordinate system is made. OGRe stays true to this definition, but takes away the complexities that come with combining tensors in different representations. This is done using an object-oriented programming approach, as detailed in the documentation.
+Tensors are abstract objects, which can be represented as multi-dimensional arrays once a choice of index configuration and coordinate system is made. OGRe stays true to this definition, but takes away the complexities that come with combining tensors in different representations. This is done using an object-oriented programming approach, taking advantage of principles such as encapsulation and class invariants to eliminate the possibility of user error.
 
 The user initially defines each tensor in OGRe using its explicit components in any single representation. Operations on this tensor are then done abstractly, without needing to specify which representation to use. Possible operations include addition of tensors, multiplication of tensor by scalar, trace, contraction, and partial and covariant derivatives.
 
@@ -39,23 +36,23 @@ Transformations between representations are done behind the scenes; all the user
 
 I initially created OGRe for use in my own research, so I made it as flexible and powerful as possible. I also wanted my students to be able to use it easily and efficiently, even if they only have minimal experience with Mathematica and/or general relativity, so I made it simple to learn and easy to use. As a result, this package is equally suitable for both experienced and novice researchers.
 
-<a id="markdown-features" name="features"></a>
 ## Features
 
 * Define coordinate systems and the transformation rules between them. Tensor components are then transformed automatically between coordinates behind the scenes as needed.
 * Each tensor is associated with a specific metric. Tensor components are then transformed automatically between different index configurations, raising and lowering indices behind the scenes as needed.
[…]

[…]
