### 2021-08-08 #3416 @bshoshany
@amelialdrew: The new release is now up, along with the modifications you requested to the paper in the `paper` branch. I believe I solved the parallelization bug, so please check and let me know if it works now or if you still get an error.

I am happy to answer any additional questions and make additional modifications as needed.
