### 2021-08-08 #3416 @amelialdrew
Hello @bshoshany, thanks for the new changes. The new version works and for me gives ```16 parallel kernels launched. CPU has 4 cores``` which is more than the default, as I think you intended. It still only speeds up from ~13s to ~8s, but I would guess this is system/version dependent, and as it does speed up I am happy to approve.
