### 2021-08-08 #3416 @bshoshany
Hi @amelialdrew, I do believe the speedup you get is exactly what I would expect you to get.

The number that counts is the number of **cores**, since that is the number of things the CPU can actually do in parallel. On my computer I have 12 cores but only 8 kernels due to license limitations, which means 4 of my 12 CPU cores are not actually being utilized. With the 8 kernels running in parallel on 8 cores, I get a speedup from ~10s to ~3.5s, which is roughly a 2.9x speedup. (One might expect to get an 8x speedup, but Mathematica's parallelization is very inefficient.)

On your computer your license allows 16 kernels, but that doesn't mean much, because you only have 4 cores, so only 4 of these kernels are actually running in parallel at any given time. Therefore the number of actual parallel kernels you have (4) is **half** the number of actual parallel kernels I have (8). You get a 1.6x speedup, which is roughly **half** the speedup I get, so that's perfectly consistent.

Thank you so much for completing the review! 🙂
