### 2021-08-30 #3416 @VivianePons
@bshoshany thanks! Can you make a tagged release of the paper / software (on Zenodo, figshare, or other) and send me the DOI? Be careful to use the same metadata (title, author) as for the paper itself
