### 2021-08-30 #3416 @bshoshany
@VivianePons I'm not sure I understand, doesn't JOSS generate its own DOI for the paper? The paper proof shows `10.21105/joss.03416` as the DOI.
