### 2021-08-30 #3416 @bshoshany
The latest version of OGRe is indeed v1.6 from 2021-08-07. I did change `paper.md` a few days ago due to @kostunin's comments, but that's not reflected in Zenodo since it only refers to the software and not the paper.

I changed the title now. Please let me know if you also need the paper to be included in the tagged release, or if just the software is enough.
