### 2021-08-30 #3416 @kostunin
> Hi @kostunin you forgot to check the "Performance" box. The box needs to be checked even if there is no performance claim.

done
