### 2021-08-31 #3416 @kyleniemeyer
Hi @bshoshany, I'm the AEIC on duty this week, doing some final checks. I just have a few minor edits for your paper before accepting:
- [ ] Could you add a bit more to your affiliation? Usually we see at least department/unit in addition to institution
- [ ] Throughout the article, you used hyphens (`-`) as em-dashes (`—`). Can you replace those with the latter? For example, in the first paragraph, "...general relativity - the field..." should be "...general relativity—the field...".
- [ ] At the end of the statement of need section, if you are including the references in the actual sentence, then I think you should use the `@author:2001` citation format rather than `[@author:2001]`. This would change (e.g.) "differential geometry in (Kobayashi & Nomizu, 1996a)" to "differential geometry in Kobayashi & Nomizu (1996a)". The other option is to remove the "in" before the references and combine, like "differential geometry (Kobayashi & Nomizu, 1996a; Kobayashi & Nomizu, 1996b)" by changing the text to `differential geometry [@Kobayashi1; @Kobayashi2]`.
