### 2021-08-31 #3416 @bshoshany
Hi @kyleniemeyer, thanks for the comments! I have applied all of your suggestions to the paper. Please let me know if any other changes are needed.
